Restrict docker build and push to develop and main branches
The build and push steps for both the webserver and robot Docker containers in the GitLab CI/CD pipeline are now only run on the develop and main branches. This change was made to ensure Docker images are only built and pushed for major branches, reducing unnecessary builds on other branches.