Refactor codebase: Split webserver into frontend and backend services
This refactor involves a separation of the Raspberry Pi webserver into distinct frontend and backend services. The Dockerfile for the webserver has been removed, numerous files have been relocated and renamed to reflect this split. The `session.py`, `webserver.py`, and `database.py` files have also been significantly modified in order to conform with this new structure. Modifications to the `.gitlab-ci.yml` file ensure that the correct directories are targeted during the build process.